+++ src/test/java/au/com/mountainpass/ryvr/testclient/RyvrTestEmbeddedServerAdminDriver.java
@@ -16,43 +16,41 @@ import cucumber.api.Scenario;
 @Component
 @Profile(value = { "integrationTest" })
 public class RyvrTestEmbeddedServerAdminDriver implements RyvrTestServerAdminDriver {
-    @Autowired
-    private JdbcTemplate currentJt;
+  @Autowired
+  private JdbcTemplate currentJt;
 
-    public final Logger LOGGER = LoggerFactory.getLogger(this.getClass());
+  public final Logger LOGGER = LoggerFactory.getLogger(this.getClass());
 
-    @Autowired
-    private RyvrsCollection ryvrsCollection;
+  @Autowired
+  private RyvrsCollection ryvrsCollection;
 
-    @Override
-    public void _after(final Scenario scenario) {
-        // nothing
-    }
+  @Override
+  public void _after(final Scenario scenario) {
+    // nothing
+  }
 
-    @Override
-    public void _before(final Scenario scenario) {
-        clearRyvrs();
-    }
+  @Override
+  public void _before(final Scenario scenario) {
+    clearRyvrs();
+  }
 
-    @Override
-    public void clearRyvrs() {
-        ryvrsCollection.clear();
-    }
+  @Override
+  public void clearRyvrs() {
+    ryvrsCollection.clear();
+  }
 
-    @Override
-    public void createDataSourceRyvr(final Map<String, String> config)
-            throws Throwable {
-        final DataSourceRyvr ryvr = new DataSourceRyvr(config.get("name"),
-                currentJt, config.get("database"), config.get("table"),
-                config.get("ordered by"),
-                Long.parseLong(config.get("page size")));
-        ryvrsCollection.addRyvr(ryvr);
+  @Override
+  public void createDataSourceRyvr(final Map<String, String> config) throws Throwable {
+    final DataSourceRyvr ryvr = new DataSourceRyvr(config.get("name"), currentJt,
+        config.get("database"), config.get("table"), config.get("ordered by"),
+        Integer.parseInt(config.get("page size")));
+    ryvrsCollection.addRyvr(ryvr);
 
-    }
+  }
 
-    @Override
-    public void ensureStarted() {
-        // nothing
-    }
+  @Override
+  public void ensureStarted() {
+    // nothing
+  }
 
 }
